refactor: 마, 사, 상, 궁 클래스를 StepPatternPiece로 추상화하여 중복 제거
